requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
Faker==37.8.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
